Comments for a few answers added.
FIrst of all, Great job!

There are a few things I want you to check.
Go through the comments, and if you want to add something, add it.
If you want to change answers, do not change them. Instead add the new answer below the comment, so that we can check the difference between old and new answer easily. Do not change the comment added also.